Remove bootloader packages
At the moment it seems more likely that the bootloader will be
updated separately from the Arch Linux installation. The main ESP
partition is too small to properly store a kernel+ramdisk in it,
which is why I've adopted another partition as additional EFI partition.

This means that the /boot partition does no longer point to the actual
ESP partition with the bootloader, so the partition with the bootloader
is not actually necessarily mounted.