archiso: Remove intel-ucode-byt-t-c0, reduce diff
intel-ucode-byt-t-c0 will not be part of the binary repository,
which makes building the ISO impossible without copying packages
manually. Also it should be up to users to configure that themselves
if they would like to use it. It's not necessary to run the ISO.

To reduce the diff, restore the original intel-ucode and amd-ucode
even though it does not actually contain any microcode for this
tablet.